Day 6 Part 2

+++ 2023/Day06/Functions.cs
@@ -14,4 +14,4 @@ internal static class Functions
     }
 }
 
-internal record Race(int TimeLimit, int BestDistance);
+internal record Race(long TimeLimit, long BestDistance);
